Edit And Add JS Files Commit

+++ server/src/controllers/zweetController.js
@@ -1,10 +1,10 @@
-import jwt, { decode } from "jsonwebtoken";
+import jwt from "jsonwebtoken";
 import Zweet from "../models/Zweet";
 import User from "../models/User";
 
 export const zweets = async (req, res) => {
     try {
-        const zweets = await Zweet.find({}).sort({ createdAt: "desc" });
+        const zweets = await Zweet.find({}).sort({ views: "desc", like: "desc" });
         return res.status(200).json(zweets);
     } catch (error) {
         return res.status(404).json({
@@ -106,7 +106,7 @@ export const zweetReaction = async (req, res) => {
                 }
             });
         };
-        const updatedZweet = await Zweet.findById(id);
+        const updatedZweet = await Zweet.findById(id).populate("owner");
         return res.status(200).json(updatedZweet);
     } catch (error) {
         return res.status(409).json({
@@ -118,17 +118,26 @@ export const zweetReaction = async (req, res) => {
 
 export const edit = async (req, res) => {
     const { id } = req.params;
+    const { token } = req.headers;
     const { title, description, hashtags } = req.body;
     try {
-        await Zweet.findByIdAndUpdate(id, {
-            title, 
-            description, 
-            hashtags: Zweet.formatHashtags(hashtags)
-        });
-        const updatedZweet = await Zweet.findById(id);
-        return res.status(200).json(updatedZweet);
+        const zweet = await Zweet.findById(id).populate("owner");
+        const decodedToken = await jwt.decode(token);
+        if (zweet.owner._id == decodedToken.user._id) {
+            await Zweet.findByIdAndUpdate(id, {
+                title, 
+                description, 
+                hashtags: Zweet.formatHashtags(hashtags)
+            });
+            const updatedZweet = await Zweet.findById(id);
+            return res.status(200).json(updatedZweet);
+        } else {
+            return res.status(400).json({
+                success: false, 
+                error: "You are not person that made this zweet."
+            });
+        };
     } catch (error) {
-        console.log(error)
         return res.status(409).json({
             success: false, 
             error: error.message
@@ -140,16 +149,24 @@ export const remove = async (req, res) => {
     const { id } = req.params;
     const { token } = req.headers;
     try {
+        const zweet = await Zweet.findById(id).populate("owner");
         const decodedToken = await jwt.decode(token);
-        await Zweet.findByIdAndDelete(id);
-        await User.findOneAndUpdate({ email: decodedToken.user.email }, {
-            $pull: {
-                zweets: id
-            }
-        });
-        return res.status(200).json({
-            success: true, 
-        });
+        if (zweet.owner._id == decodedToken.user._id) {
+            await Zweet.findByIdAndDelete(id);
+            await User.findOneAndUpdate({ email: decodedToken.user.email }, {
+                $pull: {
+                    zweets: id
+                }
+            });
+            return res.status(200).json({
+                success: true, 
+            });
+        } else {
+            return res.status(400).json({
+                success: false, 
+                error: "You are not person that made this zweet."
+            });
+        };
     } catch (error) {
         return res.status(409).json({
             success: false,
